Refactor validation approach
