--- src/Promitor.Scraper/Validation/IValidation.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Promitor.Scraper.Validation
-{
-    public interface IValidation
-    {
-        Task ValidateAsync();
-    }
-}
