+++ src/Promitor.Scraper/Validation/Interfaces/IValidationStep.cs
@@ -1,0 +1,7 @@
+﻿namespace Promitor.Scraper.Validation.Interfaces
+{
+    public interface IValidationStep
+    {
+        void Validate();
+    }
+}
