--- src/Promitor.Scraper/Validation/ScrapingScheduleValidation.cs
@@ -1,19 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Promitor.Scraper.Validation
-{
-    public class ScrapingScheduleValidation : IValidation
-    {
-        public Task ValidateAsync()
-        {
-            return Task.CompletedTask;
-        }
-    }
-    public class ScrapingSchedule2Validation : IValidation
-    {
-        public Task ValidateAsync()
-        {
-            return Task.CompletedTask;
-        }
-    }
-}
